Home
Here is were we show that bootstrap is loading fine, and all components
are ready to use.

Also, we want to show the usage of iterators with ruby, and is a good
opportunity to show a visual example. In order to show them how it
works, this adds a dynamic render of card elements in the home page.