Flash messages
After installing devise, they shows some setup suggestions, and they
mentioned to ensure to have flahs messages.

These messages are very useful to keep our user informed about their
requests, and to inform if something is succesful or everything is
wrong.

This add the flash messages in the app layout